implement "vardict" decoding
A common pattern in DBus APIs is to implement extensible
structs as a dict of dynamically typed values. The literal
translation of that is a map[K]dbus.Variant, which is akward
to work with.

However, when this type is used to implement extensible
structs, there is a single known type for a given key, so
for example in a given protocol key "foo" is always a uint32.

vardict annotations on a struct make the marshaler and unmarshaler
treat the struct as such an extensible dict, with the appropriate
keys for each struct field provided by field annotations.

Signed-off-by: David Anderson <dave@natulte.net>