Adjust API docs, clean up leftover duplicate logic.
As part of adjusting API docs, AlignDBus and SignatureDBus are
now required to return constant values. Marshaler had some caveating
that if you _really_ knew what you're doing, you could attempt
to encode with signatures/alignments that vary by value, but I
never found a use for that and it's a footgun that shouldn't be
encouraged.

Signed-off-by: David Anderson <dave@natulte.net>