Replace AlignDBus with IsDBusStruct in (Un)Marshaler.
The method is a kludge either way, but the core of what we're
trying to find out is not the value's alignment requirements
(the actual encoder/decoder functions can deal with that), but
rather specifically whether an empty array of these values
needs to be padded to 4 bytes or 8 bytes - i.e. does the
custom type behave like a DBus struct value?

The changed API answers this question directly, rather than
consume an alignment value only to immediately == 8 it.

Signed-off-by: David Anderson <dave@natulte.net>