Update TODOs.
On reflection, Marshaler/Unmarshaler will not be broken up further.
After changing AlignDBus into IsDBusStruct, there is not much to be
gained from splitting the methods up. On the other hand there is much
to lose in clarity from exposing a zillion tiny interfaces that
interact at a distance with each other and require careful reading
of documentation and source code to predict.

Signed-off-by: David Anderson <dave@natulte.net>