Fix cache concurrency bug.
Handling recursion directly in the cache meant that concurrent use of
the cache could result in one goroutine hitting the other goroutine's
recursion error. Instead, move all the recursion managmenet into the
caller, with explicit stack management.

This may result in goroutines duplicating some work when cached values
are first computed, but the computation is deterministic and cheap enough
that it's not worth the bother to singleflight them and complicate the
cache API.

Signed-off-by: David Anderson <dave@natulte.net>