Remove CallOptions, replace with context values.
Interactive authorization and autostart blocking are fairly wide
ranging policy choices that are a good fit for inclusion in the
call context. no-reply on the other hand is more of an intrinsic
property of a particular call (i.e. whether it even makes sense
to no-reply), so replace that with a OneWay method that's separate
from the blocking Call method.

Signed-off-by: David Anderson <dave@natulte.net>